Refactor: split up builtins module
The `Fregot.Eval.Builtins` module was getting rather large.  This
patch splits it up three different parts:

 -  `Fregot.Builtins.Internal` provides the core primitives for constructing
    and running builtins.
 -  The implementations move to submodules:
     *  `Fregot.Builtins.Json`
     *  `Fregot.Builtins.Regex`
     *  `Fregot.Builtins.Time`
     *  `Fregot.Builtins.Basics` (catch-all, we can break it up further)
 -  `Fregot.Builtins` which imports all of these and exports them as a
    convenient set.
