+++ lib/Fregot/Names/Renamer.hs
@@ -31,9 +31,9 @@ import qualified Data.HashSet.Extended     as HS
 import           Data.List.Extended        (splits, unsnoc)
 import           Data.Maybe                (listToMaybe, maybeToList)
 import           Data.Traversable          (for)
+import           Fregot.Builtins.Internal  (ReadyBuiltin)
 import           Fregot.Error              (Error)
 import qualified Fregot.Error              as Error
-import           Fregot.Eval.Builtins      (ReadyBuiltin)
 import           Fregot.Names
 import           Fregot.Names.Imports
 import           Fregot.Prepare.Ast        (Function (..), Imports)
